pyproject.toml
[project]
name = "openshift-ansible-cicd"
version = "0.1.0"
description = "Python CLI to drive CI/CD with OpenShift + Ansible"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12.0",
  "kubernetes>=30.1.0",
  "openshift>=0.13.2",
  "PyYAML>=6.0.2",
  "jinja2>=3.1.4",
  "ansible-runner>=2.4.0",
]

[project.scripts]
cicd = "cicd.cli:app"

[tool.ruff]
line-length = 100
